Move Limine.CFG from src/ to zealbooter/.
Since Zeal now has Limine.CFG autogenerate capability for install, it doesn't need it in the source tree. However, it can't be outright deleted from the project since the isohybrid needs it, so instead it will stay in zealbooter directory, and in the next commit it will be copied into the isohybrid temp directory from zealbooter/ during the build process.