Implement Realtek 8139 network driver
These changes were received via email on May 29 2025. Thanks Anthony! :) The body of the email was the following:

Hello,

Attached are the RTL8139 driver file and a patched Run.ZC file with the driver added.

The defines which are not used can be removed to lower the line count a bit, I just copy pasted them from other drivers as a quick reference outside of the datasheet, some of the register definitions depend on the exact model/revision of the RTL8139 family and some are entirely undocumented in the datasheet. My specific model is a TP link TF-3239DL.

Some of the comments I left were for personal reference and aren't particularly helpful, feel free to remove them as well.

Some of the magic numbers have comments next to them instead of using a definition.

I removed the cursor location symbol.

Have not tested it in QEMU (yet), but it does work on physical hardware consisting of an HP Compaq Elite 8200SFF. On an unrelated note, for whatever reason I could not get ZealOS to boot in ACPI mode on stock firmware, although TempleOS worked fine in IDE mode. Meanwhile when using Coreboot firmware I could get ZealOS to boot in ACPI mode but not TempleOS in IDE mode. Seems like the different BIOS implementations initialize something differently, but this is not an issue with ZealOS or TempleOS.

Thanks.